Patch up a few minor issues with demo + linux builds.
When I was working on the macOS build, I broke the non-Conan linux build by assuming physfs would have a pkg-config file.  No good - revert, but add some bits so we can find the physfs header too if it's not in the system include path.

Also fixes the macOS demo build to match the full build.