Reset the video settings in a more sensible manner.
This tidies up a few inconsistencies in how we were managing window state.

First of all, we only assert fullscreen vs windowed if the flag changes.

Secondly, we now flip the texture hint and rebuild the target texture when we toggle the pixel perfect option.

Finally, we now use the vid_reset_settings() function to do the initial setup of the texture, rather than just using it to fix things later.

Unfortunately, the hint is only used at texture creation time, meaning we can't just flip it on the texture (which would be possible if we were using GL directly), we do need to destroy and recreate it.